settings: stick to int64 for IntSetting
it is up to a caller to decide if they want to convert to/from `int`,
but since we use an atomic int64 internally, let’s expose it in API.

+++ pkg/sql/set.go
@@ -160,7 +160,7 @@ func (p *planner) toSettingString(
 	case *settings.IntSetting:
 		return typeCheckAndParse(parser.TypeInt, func(d parser.Datum) (string, error) {
 			if i, ok := d.(*parser.DInt); ok {
-				return settings.EncodeInt(int(*i)), nil
+				return settings.EncodeInt(int64(*i)), nil
 			}
 			return "", errors.Errorf("cannot use %s %T value for int setting", d.ResolvedType(), d)
 		})
